pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "matt-james-platform"
version = "0.1.0"
description = "Core multi-tenant platform foundation API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.0",
  "structlog>=25.0.0",
  "python-dotenv>=1.0.0",
  "PyJWT>=2.9.0",
  "psycopg[binary]>=3.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.28.0",
  "ruff>=0.8.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.28.0",
  "ruff>=0.8.0",
]

[tool.uv]
default-groups = ["dev"]

[tool.hatch.build.targets.wheel]
packages = ["api"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["api", "tests", "scripts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
